Adds ruby support
[noissue]

+++ generate.sh
@@ -7,22 +7,36 @@ if [ $# -eq 0 ]; then
 fi
 
 curl -o api.json http://localhost:24817/pulp/api/v3/docs/api.json?plugin=$1
-docker run --rm -v ${PWD}:/local swaggerapi/swagger-codegen-cli generate \
-    -i /local/api.json \
-    -l python \
-    -o /local/$1-client \
-    -DpackageName=pulpcore.client.$1 \
-    -DprojectName=$1-client \
-    -DpackageVersion=$2
-cp python/__init__.py $1-client/pulpcore/
-cp python/__init__.py $1-client/pulpcore/client
 
-# There is a bug in swagger-codegen. When using package names within a namespace it creates the package
-# accross 2 different directories. We move everything back into place here.
-cp $1-client/pulpcore.client.$1/* $1-client/pulpcore/client/$1/
-cp $1-client/pulpcore.client.$1/api/* $1-client/pulpcore/client/$1/api/
-cp $1-client/pulpcore.client.$1/models/* $1-client/pulpcore/client/$1/models/
-# Then remove the wrong directory
-rm -rf $1-client/pulpcore.client.$1
+if [ $3 = 'python' ]
+then
+    docker run --rm -v ${PWD}:/local swaggerapi/swagger-codegen-cli generate \
+        -i /local/api.json \
+        -l python \
+        -o /local/$1-client \
+        -DpackageName=pulpcore.client.$1 \
+        -DprojectName=$1-client \
+        -DpackageVersion=$2
+    cp python/__init__.py $1-client/pulpcore/
+    cp python/__init__.py $1-client/pulpcore/client
+    # There is a bug in swagger-codegen. When using package names within a namespace it creates the package
+    # accross 2 different directories. We move everything back into place here.
+    cp $1-client/pulpcore.client.$1/* $1-client/pulpcore/client/$1/
+    cp $1-client/pulpcore.client.$1/api/* $1-client/pulpcore/client/$1/api/
+    cp $1-client/pulpcore.client.$1/models/* $1-client/pulpcore/client/$1/models/
+    # Then remove the wrong directory
+    rm -rf $1-client/pulpcore.client.$1
+fi
+if [ $3 = 'ruby' ]
+then
+    curl -o api.json http://localhost:24817/pulp/api/v3/docs/api.json?plugin=$1
+    docker run --rm -v ${PWD}:/local swaggerapi/swagger-codegen-cli generate \
+        -i /local/api.json \
+        -l ruby \
+        -o /local/$1-client \
+        -DgemName=$1_client \
+        -DgemLicense="GPLv2" \
+        -DgemVersion=$2
+fi
 
 rm api.json
